Rename package to match 2.3 version
